import own stylus files, rather than some random file in the stylus path that happens to have the same name
"lib/index.style" was importing the wrong .styl file if another file
of the same name happens to exist in the path.

For example, if there is a "normalize.styl" in the stylus path, it is
used instead of the correct file "nib/lib/normalize/index.styl".

This commit changes index.styl to use imports relative to the current
directory, which fixes this issue.